fix(prediction): address review findings (Go winner determinism + 3 more)
From a cross-language review of this session's changes:

- limitless: the resolved-market winner was `i === winningOutcomeIndex` where i
  iterates Object.keys(tokens) — Go/Java randomize map iteration, so ~50% of the
  time the wrong outcome was flagged winner (JS/Python are insertion-ordered so
  it passed live). Map the leg to its canonical index by label (yes=0, no=1)
  instead of loop position. Now deterministic across all languages.
- polymarket: a market that is closed-for-trading but not yet UMA-resolved has
  fractional outcomePrices; the code reported settleFraction = the mid (e.g. 0.5)
  as if final. Only set winner/settleFraction when a decisive price exists
  (>=0.99 or <=0.01); leave undefined otherwise.
- kalshi editOrder: validate price + amount BEFORE cancelling the old order, so a
  bad edit (kalshi is limit-only) doesn't leave the user with the order cancelled
  and an exception instead of a replacement.
- C# PredictionSettlement.timestamp: Int64? not double?, matching every sibling
  struct + Go (*int64) / Python (Int) / TS (Int).

Verified: limitless (winningOutcomeIndex=1 -> NO winner) + polymarket (closed-
unsettled -> no winner/settleFraction) via mocks; live re-check confirms Trump-
2024 (prices 1/0) still winner=true/settleFraction=1 and myriad categorical
intact; kalshi editOrder throws before cancel on missing price. Full offline
matrix green JS/C#/Go; C#/Go build; gofmt clean.

+++ cs/ccxt/base/PredictionTypes.cs
@@ -392,7 +392,7 @@ public struct PredictionTickers
 public struct PredictionSettlement
 {
     public string? id;
-    public double? timestamp;
+    public Int64? timestamp;
     public string? datetime;
     public string? outcome;
     public string? outcomeId;
